Mysql db

+++ controller/post.mjs
@@ -1,6 +1,6 @@
 import * as postRepository from "../data/post.mjs";
 
-//모든 포스트/ 해당 아이디에 대한 포스트를 가져오는 함수
+// 모든 포스트 / 해당 아이디에 대한 포스트를 가져오는 함수
 export async function getPosts(req, res, next) {
   const userid = req.query.userid;
   const data = await (userid
@@ -9,7 +9,7 @@ export async function getPosts(req, res, next) {
   res.status(200).json(data);
 }
 
-//id를 받아 하나의 포스트를 가져오는 함수
+// id를 받아 하나의 포스트를 가져오는 함수
 export async function getPost(req, res, next) {
   const id = req.params.id;
   const post = await postRepository.getById(id);
@@ -22,12 +22,13 @@ export async function getPost(req, res, next) {
 
 // 포스트를 생성하는 함수
 export async function createPost(req, res, next) {
-  const { userid, name, text } = req.params.body;
-  const post = await postRepository.create(userid, name, text);
-  res.status(201).json(post);
+  const { text } = req.body;
+  console.log("req.useridx: ", req.useridx);
+  const posts = await postRepository.create(text, req.useridx);
+  res.status(201).json(posts);
 }
 
-//포스트 변경
+// 포스트를 변경하는 함수
 export async function updatePost(req, res, next) {
   const id = req.params.id;
   const text = req.body.text;
@@ -39,8 +40,17 @@ export async function updatePost(req, res, next) {
   }
 }
 
+// 포스트를 삭제하는 함수
 export async function deletePost(req, res, next) {
   const id = req.params.id;
+  const post = await postRepository.getById(id);
+  if (!post) {
+    return res.status(404).json({ message: `${id}의 포스트가 없습니다` });
+  }
+  if (post.useridx !== req.useridx) {
+    return res.sendStatus(403);
+  }
   await postRepository.remove(id);
+  console.log("삭제됨!");
   res.sendStatus(204);
 }
